// ManagmentServer/API/UI/src/Components/Monitor/AdapterList/__generated__/AdapterListDataFragment.graphql.ts
export type AdapterListDataFragment$data = {
  readonly adapters: {
    readonly __id: string;
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly id: string;
        readonly " $fragmentSpreads": FragmentRefs<"AdapterListItemDataFragment">;
      } | null;
    }> | null;
    readonly pageInfo: {
      readonly endCursor: string | null;
      readonly hasNextPage: boolean;
      readonly hasPreviousPage: boolean;
      readonly startCursor: string | null;
    };
  } | null;
  readonly " $fragmentType": "AdapterListDataFragment";
};
export type AdapterListDataFragment$key = {
  readonly " $data"?: AdapterListDataFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"AdapterListDataFragment">;
};